Fix compile, Add SM8521 to list
