+++ src/engine/dispatchContainer.cpp
@@ -75,6 +75,7 @@
 #include "platform/vb.h"
 #include "platform/k007232.h"
 #include "platform/ga20.h"
+#include "platform/sm8521.h"
 #include "platform/pcmdac.h"
 #include "platform/dummy.h"
 #include "../ta-log.h"
@@ -488,6 +489,9 @@ void DivDispatchContainer::init(DivSystem sys, DivEngine* eng, int chanCount, do
       break;
     case DIV_SYSTEM_GA20:
       dispatch=new DivPlatformGA20;
+      break;
+    case DIV_SYSTEM_SM8521:
+      dispatch=new DivPlatformSM8521;
       break;
     case DIV_SYSTEM_PCM_DAC:
       dispatch=new DivPlatformPCMDAC;
