+++ src/engine/song.h
@@ -124,7 +124,8 @@ enum DivSystem {
   DIV_SYSTEM_YM2610_CSM,
   DIV_SYSTEM_YM2610B_CSM,
   DIV_SYSTEM_YM2203_CSM,
-  DIV_SYSTEM_YM2608_CSM
+  DIV_SYSTEM_YM2608_CSM,
+  DIV_SYSTEM_SM8521
 };
 
 struct DivGroovePattern {
